refactor: 도메인 및 컨트롤러 패키지 구조 세분화
