+++ src/main/java/blackjack/core/BlackjackGame.java
@@ -6,13 +6,13 @@ import blackjack.dto.InitialDealDtos;
 import blackjack.dto.ParticipantCardsDto;
 import blackjack.dto.ParticipantScoreDto;
 import blackjack.dto.GameResultDtos;
-import blackjack.model.CardsGenerator;
-import blackjack.model.Dealer;
-import blackjack.model.Deck;
-import blackjack.model.GameResult;
-import blackjack.model.Participant;
-import blackjack.model.Participants;
-import blackjack.model.Player;
+import blackjack.domain.card.CardsGenerator;
+import blackjack.domain.participant.Dealer;
+import blackjack.domain.card.Deck;
+import blackjack.domain.result.GameResult;
+import blackjack.domain.participant.Participant;
+import blackjack.domain.participant.Participants;
+import blackjack.domain.participant.Player;
 import blackjack.view.BlackjackView;
 import java.util.List;
