+++ src/main/java/blackjack/domain/card/Hand.java
@@ -1,5 +1,6 @@
-package blackjack.model;
+package blackjack.domain.card;
 
+import blackjack.domain.result.Score;
 import java.util.ArrayList;
 import java.util.List;
