+++ src/main/java/blackjack/domain/card/ShuffledCardsGenerator.java
@@ -1,4 +1,4 @@
-package blackjack.model;
+package blackjack.domain.card;
 
 import java.util.ArrayList;
 import java.util.Arrays;
