+++ src/main/java/blackjack/domain/card/Suit.java
@@ -1,4 +1,4 @@
-package blackjack.model;
+package blackjack.domain.card;
 
 public enum Suit {
     DIAMOND, CLOVER, SPADE, HEART
