+++ src/main/java/blackjack/domain/participant/Player.java
@@ -1,4 +1,6 @@
-package blackjack.model;
+package blackjack.domain.participant;
+
+import blackjack.domain.card.Hand;
 
 public class Player extends Participant {
