+++ src/main/java/blackjack/domain/result/GameResult.java
@@ -1,4 +1,4 @@
-package blackjack.model;
+package blackjack.domain.result;
 
 public enum GameResult {
     PLAYER_WIN, DEALER_WIN, PUSH
