+++ src/test/java/blackjack/domain/participant/DealerTest.java
@@ -1,7 +1,12 @@
-package blackjack.model;
+package blackjack.domain.participant;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
+import blackjack.domain.card.Card;
+import blackjack.domain.card.Hand;
+import blackjack.domain.card.Rank;
+import blackjack.domain.card.Suit;
+import blackjack.domain.result.GameResult;
 import java.util.List;
 import org.junit.jupiter.api.Test;
