+++ src/test/java/blackjack/domain/participant/PlayerTest.java
@@ -1,8 +1,9 @@
-package blackjack.model;
+package blackjack.domain.participant;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
+import blackjack.domain.card.Hand;
 import org.junit.jupiter.api.Test;
 
 class PlayerTest {
