+++ src/test/java/blackjack/domain/result/ScoreTest.java
@@ -1,7 +1,10 @@
-package blackjack.model;
+package blackjack.domain.result;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import blackjack.domain.card.Card;
+import blackjack.domain.card.Rank;
+import blackjack.domain.card.Suit;
 import java.util.List;
 import org.assertj.core.api.AssertionsForClassTypes;
 import org.junit.jupiter.api.Test;
